Images in a grid gallery.

gridGallery produces a gridGallery object.

gridGallery(nodes, name = NULL, label = NULL, info = NULL, image = NULL,
            main = NULL, note = NULL, help = NULL,
            language = c("en", "es", "ca"), dir = NULL)

Arguments

nodes
a data frame with at least three vectors of names, start and end.

name
name of the vector with names in the nodes data frame.

label
name of the vector with labels in the nodes data frame.

main
upper title of the graph.

note
lower title of the graph.

info
name of the vector with information to display in a panel in the nodes data frame.

language
a character vector (es=spanish; en=english; ca=catalan).

help
help text of the graph.

image
name of the vector with image files in the nodes data frame.

dir
a "character" string representing the directory where the web files will be saved.

Value

Object of class gridGallery.

Author

Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See https://sociocav.usal.es/blog/modesto-escobar/

Examples

data("finches")
finches$species<-system.file("extdata", finches$species,
        package="netCoin") # copy path to the species field
gridGallery(finches, image="species", main="Species in Galapagos Islands",
        note="Data source: Sanderson (2000)")

#> Title: Species in Galapagos Islands 
#> 
#> Nodes(13):
#>                   name frequency     type
#>  Geospiza magnirostris        14 Geospiza
#>        Geospiza fortis        13 Geospiza
#>    Geospiza fuliginosa        14 Geospiza
#>    Geospiza difficilis        10 Geospiza
#>      Geospiza scandens        12 Geospiza
#>   Geospiza conirostris         2 Geospiza
#>                                                                                                    species
#>  C:/Users/Modesto/AppData/Local/Temp/Rtmp0ufTuH/temp_libpath5dd8b3a1ef7/netCoin/extdata/g.Magnirostris.png
#>        C:/Users/Modesto/AppData/Local/Temp/Rtmp0ufTuH/temp_libpath5dd8b3a1ef7/netCoin/extdata/g.Fortis.png
#>    C:/Users/Modesto/AppData/Local/Temp/Rtmp0ufTuH/temp_libpath5dd8b3a1ef7/netCoin/extdata/g.Fuliginosa.png
#>    C:/Users/Modesto/AppData/Local/Temp/Rtmp0ufTuH/temp_libpath5dd8b3a1ef7/netCoin/extdata/g.Difficilis.png
#>      C:/Users/Modesto/AppData/Local/Temp/Rtmp0ufTuH/temp_libpath5dd8b3a1ef7/netCoin/extdata/g.Scandens.png
#>   C:/Users/Modesto/AppData/Local/Temp/Rtmp0ufTuH/temp_libpath5dd8b3a1ef7/netCoin/extdata/g.Conirostris.png
#> ...
#> 
#> Data source: Sanderson (2000)