Produce interactive multi graphs.

multigraphCreate produce an interactive multi graph.

multigraphCreate(..., mode = c("default","parallel","frame"),
                 frame = 0, speed = 50, dir = "MultiGraph", show = TRUE)

Arguments

...
rD3plot graphs (network_rd3, barplot_rd3, timeplot_rd3) objects or html "directories".

mode
a string specifying the displaying mode. The "default" displays graphs one by one, "parallel" splits screen and "frame" allows dinamic graphs in time.

frame
number of frame to start a dynamic network.

speed
a percentage for frame speed in dynamic networks.

dir
a "character" string representing the directory where the graph will be saved.

show
a logical value true if the graph is to be shown. Default = TRUE.

Author

Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See https://sociocav.usal.es/blog/modesto-escobar/

Value

The function creates a folder in your computer with an HTML document named index.html which contains the graph. This file can be directly opened with your browser.

Examples

if (FALSE) {
# A character column (with separator)
frame <- data.frame(A = c("Man; Women", "Women; Women",
                        "Man; Man", "Undet.; Women; Man"))
data <- dichotomize(frame, "A", sep = "; ")[2:4]
C <- coin(data) # coincidence matrix
N <- asNodes(C) # node data frame
E <- edgeList(C,c("frequency","expected","haberman")) # edge data frame
bC<- barCoin(data,dichotomies="_all") # barCoin object
cC<- barCoin(data,dichotomies="_all",expected=TRUE) # barCoin object
nC<- netCoin(N,E) # netCoin object
multigraphCreate("Bar graph" = bC,
                  "Conditional bar graph" = cC, 
                  "Net graph"=nC, 
                  dir="./example") # See ./example/index.html file
}