Networked coincidences.

timeCoin produces a timeCoin object.

timeCoin(periods, name = "name", start = "start", end = "end", group = NULL,
         text = NULL, main = NULL, note = NULL, info = NULL, 
         events = NULL, eventNames = "name", eventPeriod = "period",
         eventTime = "date", eventColor = NULL, eventShape = NULL,
         cex = 1, language = c("en","es","ca"), dir = NULL)

Arguments

periods
a data frame with at least three vectors of name, start and end of the periods.

name
name of the vector with names in the periods data frame.

start
name of the vector with starts in the periods data frame.

end
name of the vector with ends in the periods data frame.

group
name of the vector with groups in the periods data frame.

text
name of the vector with html text in the periods data frame.

main
upper title of the graph.

note
lower title of the graph.

info
name of the vector with information to display in a panel in the periods data frame.

events
a data frame of events included into the periods with three columns: event name, periodParent and eventTime

eventNames
name of the vector with names in the events data frame.

eventPeriod
name of the vector with period names in the events data frame.

eventTime
name of the vector with time points in the events data frame.

eventColor
name of the vector with color criteria in the events data frame.

eventShape
name of the vector with shape criteria in the events data frame.

cex
number indicating the amount by which plotting text should be scaled relative to the default. Default = 1.

language
a character vector (es=spanish; en=english; ca=catalan).

dir
a "character" string representing the directory where the web files will be saved.

Value

Object of class timeCoin.

Author

Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See https://sociocav.usal.es/blog/modesto-escobar/

Examples

# Database of the classical sociologists.
# Only periods
data(sociologists)
timeCoin(sociologists, "name", "birth", "death", "school")

#> 
#> Periods(16):
#>                   name birth death birth_place birth_country death_place
#>          Auguste Comte  1798  1857 Montpellier       Francia       Paris
#>              Karl Marx  1818  1883       Trier      Alemania      London
#>        Herbert Spencer  1820  1903       Derby   Reino Unido    Brighton
#>       Friedrich Engels  1820  1895      Barmen      Alemania      London
#>  William Graham Sumner  1840  1910  New Jersey          EEUU Connecticut
#>        Vilfredo Pareto  1848  1923       Paris       Francia     Céligny
#>  death_country                             label generation            school
#>        Francia         Auguste Comte (1798-1857)  1775-1799     a) Positivism
#>    Reino Unido             Karl Marx (1818-1883)  1800-1824  b) Social action
#>    Reino Unido       Herbert Spencer (1820-1903)  1800-1824   c) Evolutionism
#>    Reino Unido      Friedrich Engels (1820-1895)  1800-1824  b) Social action
#>           EEUU William Graham Sumner (1840-1910)  1825-1850   c) Evolutionism
#>          Suiza       Vilfredo Pareto (1848-1923)  1825-1850 d) Socioeconomics
#>       picture
#>    Q12718.png
#>     Q9061.png
#>   Q144535.png
#>    Q34787.png
#>  Q1356390.png
#>    Q11031.png
#> ...
#> 

# Periods and events
data(works)
timeCoin(sociologists, "name", "birth", "death", "school",
         events=works, eventNames="label", eventPeriod="author", eventTime="date")

#> 
#> Periods(16):
#>                   name birth death birth_place birth_country death_place
#>          Auguste Comte  1798  1857 Montpellier       Francia       Paris
#>              Karl Marx  1818  1883       Trier      Alemania      London
#>        Herbert Spencer  1820  1903       Derby   Reino Unido    Brighton
#>       Friedrich Engels  1820  1895      Barmen      Alemania      London
#>  William Graham Sumner  1840  1910  New Jersey          EEUU Connecticut
#>        Vilfredo Pareto  1848  1923       Paris       Francia     Céligny
#>  death_country                             label generation            school
#>        Francia         Auguste Comte (1798-1857)  1775-1799     a) Positivism
#>    Reino Unido             Karl Marx (1818-1883)  1800-1824  b) Social action
#>    Reino Unido       Herbert Spencer (1820-1903)  1800-1824   c) Evolutionism
#>    Reino Unido      Friedrich Engels (1820-1895)  1800-1824  b) Social action
#>           EEUU William Graham Sumner (1840-1910)  1825-1850   c) Evolutionism
#>          Suiza       Vilfredo Pareto (1848-1923)  1825-1850 d) Socioeconomics
#>       picture
#>    Q12718.png
#>     Q9061.png
#>   Q144535.png
#>    Q34787.png
#>  Q1356390.png
#>    Q11031.png
#> ...
#> 
#> Events(54):
#>              name  label                                      works date
#>     Auguste Comte  Comte              Cours de philosophie positive 1842
#>         Karl Marx   Marx                        Die heilige Familie 1845
#>  Friedrich Engels Engels Die Lage der arbeitenden Klasse in England 1845
#>  Friedrich Engels Engels                        Die heilige Familie 1845
#>         Karl Marx   Marx                     Die deutsche ideologie 1846
#>  Friedrich Engels Engels                     Die deutsche Ideologie 1846
#> ...
#>